some final fixes in err handling

+++ server/controllers/posts.js
@@ -3,10 +3,10 @@ const Post = require("../models/posts");
 const { UnAuthorizedAccess, BadRequest, NotFoundError } = require("../errors");
 
 const createPosts = async (req, res) => {
-  req.body.author = req.user.userId;
-  const { title, content, author } = req.body;
+  req.body.postedBy = req.user.userId;
+  const { title, content, postedBy } = req.body;
 
-  if (!author) {
+  if (!postedBy) {
     throw new UnAuthorizedAccess("unauthorized access");
   }
 
@@ -14,7 +14,7 @@ const createPosts = async (req, res) => {
     throw new BadRequest("title and content cannot be empty");
   }
 
-  const post = await Post.create({ ...req.body, author });
+  const post = await Post.create({ ...req.body, postedBy });
   res.status(StatusCodes.OK).json({
     success: true,
     post,
@@ -22,7 +22,8 @@ const createPosts = async (req, res) => {
 };
 
 const getPosts = async (req, res) => {
-  const posts = await Post.find({});
+  const { userId } = req.user;
+  const posts = await Post.find({ postedBy: userId }).sort("createdAt");
 
   res.status(StatusCodes.OK).json({
     success: true,
@@ -31,12 +32,19 @@ const getPosts = async (req, res) => {
 };
 
 const updatePost = async (req, res) => {
-  const { id: postId } = req.params;
+  const {
+    params: { id: postId },
+    user: { userId },
+  } = req;
 
-  const post = await Post.findOneAndUpdate({ _id: postId }, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  const post = await Post.findOneAndUpdate(
+    { _id: postId, postedBy: userId },
+    req.body,
+    {
+      new: true,
+      runValidators: true,
+    }
+  );
 
   if (!post) {
     throw new NotFoundError(`no post with id: ${postId}`);
@@ -49,9 +57,12 @@ const updatePost = async (req, res) => {
 };
 
 const getPostById = async (req, res) => {
-  const { id: postId } = req.params;
+  const {
+    params: { id: postId },
+    user: { userId },
+  } = req;
 
-  const post = await Post.findOne({ _id: postId });
+  const post = await Post.findOne({ _id: postId, postedBy: userId });
 
   if (!post) {
     throw new NotFoundError(`no post with id: ${postId}`);
@@ -64,9 +75,12 @@ const getPostById = async (req, res) => {
 };
 
 const deletePost = async (req, res) => {
-  const { id: postId } = req.params;
+  const {
+    params: { id: postId },
+    user: { userId },
+  } = req;
 
-  const post = await Post.deleteOne({ _id: postId });
+  const post = await Post.deleteOne({ _id: postId, postedBy: userId });
 
   if (!post) {
     throw new NotFoundError(`no post with id: ${postId}`);
